MainTabsScreen [nfc]: Rename, from `MainTabs`.
Now, this matches the various other `*Screen` components we use in
`AppNavigator`.

Also, make the screen name ('main' -> 'main-tabs') match the
component name more closely.
