+++ src/nav/globalTypes.js
@@ -3,7 +3,7 @@
 import type { AppNavigatorParamList } from './AppNavigator';
 import type { SharingNavigatorParamList } from '../sharing/SharingScreen';
 import type { StreamTabsNavigatorParamList } from '../main/StreamTabsScreen';
-import type { MainTabsNavigatorParamList } from '../main/MainTabs';
+import type { MainTabsNavigatorParamList } from '../main/MainTabsScreen';
 
 export type GlobalParamList = {|
   ...AppNavigatorParamList,
